+++ src/components/Paper/Paper.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import { GradeBand } from '../../enum/gradeband';
+import { Color } from '../../utils/colors';
 
 export type PaperProps = {
   children: React.ReactNode;
   element?: 'section' | 'article' | 'main' | 'aside' | 'div';
-  elevation?: 0 | 1 | 2;
+  elevation?: -1 | 0 | 1 | 2 | 3;
   roundedCorner?: boolean;
+  backgroundColor?: Color;
   className?: string;
   dataTestId?: string;
   gradeBand?: GradeBand;
@@ -16,6 +18,7 @@ export const Paper: React.FC<PaperProps> = ({
   element: Component = 'div',
   elevation,
   roundedCorner = false,
+  backgroundColor = '--connect__white',
   className,
   dataTestId,
 }) => {
@@ -28,8 +31,16 @@ export const Paper: React.FC<PaperProps> = ({
     .filter(Boolean)
     .join(' ');
 
+  let paperProps: React.CSSProperties = {};
+
+  if (backgroundColor) {
+    paperProps = {
+      '--connect__paper-bg': `var(--connect__${backgroundColor})`,
+    } as React.CSSProperties;
+  }
+
   return (
-    <Component className={paperClassName} data-testid={dataTestId}>
+    <Component className={paperClassName} data-testid={dataTestId} style={paperProps}>
       {children}
     </Component>
   );